protection, folders

+++ backend/src/chats/services.py
@@ -4,15 +4,16 @@ import logging
 
 from fastapi import HTTPException, status
 from sqlalchemy.ext.asyncio import AsyncSession
-from sqlalchemy import select
+from sqlalchemy import select, delete
 from sqlalchemy.orm import selectinload
 
 from src.utils import save_to_db
 from src.auth.models import UserModel
-from src.auth.services import get_user_by_username_or_email
-from src.chats.schemas import CreateChatSchema, ChatSchema
+from src.chats.schemas import CreateChatSchema
 from src.chats.models import ChatModel, UserChatAssociationModel
 from src.chats.enums import ChatType
+from src.folders.models import FolderChatAssociationModel, FolderModel
+from src.folders.enums import FolderType
 
 logger = logging.getLogger(__name__)
 
@@ -30,6 +31,7 @@ async def get_chat_or_404(db: AsyncSession, chat_uuid: UUID) -> ChatModel:
     return chat
 
 async def get_user_or_404(db: AsyncSession, username: str) -> ChatModel:
+    from src.auth.services import get_user_by_username_or_email
     user = await get_user_by_username_or_email(db, username)
 
     if user is None:
@@ -37,33 +39,12 @@ async def get_user_or_404(db: AsyncSession, username: str) -> ChatModel:
     
     return user
 
-async def get_chat_list(db: AsyncSession, current_user: UserModel) -> list[ChatSchema]:
-    result = await db.execute(
-        select(ChatModel)
-        .join(ChatModel.user_associations)
-        .where(UserChatAssociationModel.user_id == current_user.id)
-        .options(selectinload(ChatModel.user_associations))
-    )
-    chats = result.scalars().all()
-    
-    schemas = []
-    for chat in chats:
-        # goes trough all users
-        for assoc in chat.user_associations:
-            if assoc.user_id == current_user.id:
-                schemas.append(ChatSchema.model_validate({
-                    **chat.__dict__,
-                    'is_pinned': assoc.is_pinned
-                }))
-                break
-
-    return schemas
-
 async def create_chat_in_db(
     db: AsyncSession,
     current_user: UserModel,
     chat_info: CreateChatSchema
 ) -> ChatModel:
+    from src.folders.services import get_all_folders
     chat = ChatModel(
         chat_type=chat_info.chat_type,
         name=chat_info.name,
@@ -77,13 +58,45 @@ async def create_chat_in_db(
     )
     db.add(chat_assoc)
 
+    folders = await get_all_folders(db, current_user)
+    all_folder = None
+    chats_folder = None
+    groups_folder = None
+    for folder in folders:
+        match folder.folder_type:
+            case FolderType.ALL:
+                all_folder = folder
+            case FolderType.CHATS:
+                chats_folder = folder
+            case FolderType.GROUPS:
+                groups_folder = folder
+    
+    all_folder_assoc = FolderChatAssociationModel(folder=all_folder, chat=chat)
+    db.add(all_folder_assoc)
+
     if chat_info.chat_type == ChatType.NORMAL:
+        chats_folder_assoc = FolderChatAssociationModel(folder=chats_folder, chat=chat)
         other_user = await get_user_or_404(db, chat_info.name)
+
+        folders = await get_all_folders(db, other_user)
+        for folder in folders:
+            match folder.folder_type:
+                case FolderType.ALL:
+                    all_folder = folder
+                case FolderType.CHATS:
+                    chats_folder = folder
+
+        other_all_folder_assoc = FolderChatAssociationModel(folder=all_folder, chat=chat)
+        other_chats_folder_assoc = FolderChatAssociationModel(folder=chats_folder, chat=chat)
+
         other_chat_assoc = UserChatAssociationModel(
             user=other_user,
             chat=chat
         )
-        db.add(other_chat_assoc)
+        db.add_all(other_chat_assoc, chats_folder_assoc, other_chats_folder_assoc, other_all_folder_assoc)
+    elif chat_info.chat_type == ChatType.GROUP:
+        group_folder_assoc = FolderChatAssociationModel(folder=groups_folder, chat=chat)
+        db.add(group_folder_assoc)
     
     await db.commit()
     logger.info(f"Chat '{chat.name}' was created by '{current_user.username}'")
@@ -91,10 +104,18 @@ async def create_chat_in_db(
 
 async def delete_chat_in_db(
     db: AsyncSession,
+    user: UserModel,
     chat_uuid: UUID
 ) -> list[UserModel]:
     chat = await get_chat_or_404(db, chat_uuid)
     users = [assoc.user for assoc in chat.user_associations]
+
+    if user not in users:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail='You must be in the group, if you want delete it'
+        )
+
     await db.delete(chat)
     await db.commit()
     logger.info(f"Chat '{chat.name}' was deleted'")
@@ -106,53 +127,60 @@ async def quit_group_in_db(
     chat_uuid: UUID
 ) -> None:
     chat = await get_chat_or_404(db, chat_uuid)
-    
+
     # goes trough all users
     for assoc in chat.user_associations:
         if assoc.user_id == current_user.id:
             await db.delete(assoc)
-            await db.commit()
             logger.info(f"'{current_user.username}' quitted chat '{chat.name}'")
             break
+    
+    stmt = (
+        delete(FolderChatAssociationModel)
+        .where(
+            FolderChatAssociationModel.chat_id == chat.id,
+            FolderChatAssociationModel.folder_id.in_(
+                select(FolderModel.id).where(FolderModel.user_id == current_user.id)
+            )
+        )
+    )
+    await db.execute(stmt)
+    await db.commit()
 
 async def add_user_to_group_in_db(
     db: AsyncSession,
     group_uuid: UUID,
-    username: str
+    username: str,
+    current_user: UserModel,
 ) -> UserModel:
+    from src.folders.services import get_all_folders
     # two requests 'parallel' are faster
     chat, user = await asyncio.gather(
         get_chat_or_404(db, group_uuid),
         get_user_or_404(db, username)
     )
     
-    chat_association = UserChatAssociationModel(
-        user=user,
-        chat=chat
-    )
+    if current_user not in chat.user_associations:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail='You must be in the group, if you want add someone'
+        )
+    
+    chat_association = UserChatAssociationModel(user=user, chat=chat)
 
-    await save_to_db(db, [chat_association])
+    folders = await get_all_folders(db, user)
+    all_folder = None
+    group_folder = None
+    for folder in folders:
+        match folder.folder_type:
+            case FolderType.ALL:
+                all_folder = folder
+            case FolderType.GROUPS:
+                group_folder = folder
+
+    all_folder_assoc = FolderChatAssociationModel(folder=all_folder, chat=chat)
+    group_folder_assoc = FolderChatAssociationModel(folder=group_folder, chat=chat)
+
+    await save_to_db(db, [chat_association, all_folder_assoc, group_folder_assoc])
     logger.info(f"'{user.username}' joined chat '{chat.name}'")
     return user
-
-# True -> chat was pinned up
-# False -> chat was unpinned
-async def pin_chat_in_db(
-    db: AsyncSession,
-    current_user: UserModel,
-    chat_uuid: UUID
-) -> bool:
-    result = await db.execute(
-        select(UserChatAssociationModel)
-        .join(ChatModel)
-        .where(ChatModel.uuid == chat_uuid)
-        .where(UserChatAssociationModel.user_id == current_user.id)
-    )
-    assoc = result.scalar_one_or_none()
-
-    if assoc is None:
-        raise HTTPException(status.HTTP_404_NOT_FOUND, 'Association not found')
-    
-    assoc.is_pinned = not assoc.is_pinned
-    await db.commit()
-    return assoc.is_pinned
